+++ lucet-runtime/tests/start.rs
@@ -1,3 +1,4 @@
-use lucet_runtime_tests::start_tests;
+use lucet_runtime_tests::{start_common_defs, start_tests};
 
+start_common_defs!();
 start_tests!(lucet_runtime::MmapRegion);